% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmtr.R
\name{sc_apply_intervention}
\alias{sc_apply_intervention}
\title{Apply a council intervention to WMTR parameters}
\usage{
sc_apply_intervention(
  params,
  param,
  direction = "increase",
  magnitude = "small"
)
}
\arguments{
\item{params}{A \code{scelo_wmtr_params}.}

\item{param}{One of \link{SC_INTERVENTION_PARAMS}.}

\item{direction}{\code{"increase"} or \code{"decrease"}.}

\item{magnitude}{\code{"small"} or \code{"large"}.}
}
\value{
The modified \code{scelo_wmtr_params}.
}
\description{
Shift one parameter by 0.07 (\code{small}) or 0.20 (\code{large}) within \verb{[0, 1]},
or step the shock environment one level up or down.
}
\examples{
sc_apply_intervention(sc_wmtr_params(), "pFamily", "increase", "large")$pFamily
}
