% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{sc_corr}
\alias{sc_corr}
\title{Correlation matrix of the numeric columns}
\usage{
sc_corr(df, method = "pearson", min_abs = 0)
}
\arguments{
\item{df}{A data frame.}

\item{method}{pearson / spearman / kendall.}

\item{min_abs}{Blank cells weaker than this.}
}
\description{
Correlation matrix of the numeric columns
}
