% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{sc_df}
\alias{sc_df}
\title{Plain data frame}
\usage{
sc_df(x)
}
\arguments{
\item{x}{A \code{scelo_table}.}
}
\value{
A data.frame.
}
\description{
Drop the Scelo attributes and return an ordinary data.frame.
}
