% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmtr.R
\name{SC_INTERVENTION_PARAMS}
\alias{SC_INTERVENTION_PARAMS}
\title{The parameters a council intervention may shift}
\format{
A character vector.
}
\usage{
SC_INTERVENTION_PARAMS
}
\description{
The parameters a council intervention may shift
}
