% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{sc_life_expectancy}
\alias{sc_life_expectancy}
\title{Life expectancy from a qx vector}
\usage{
sc_life_expectancy(q, curtate = FALSE)
}
\arguments{
\item{q}{Numeric vector of qx.}

\item{curtate}{Curtate (\code{TRUE}) or complete expectancy.}
}
\value{
A number.
}
\description{
Complete (ex, uniform deaths: curtate + 1/2) or curtate life expectancy
from the qx of one life.
}
\examples{
sc_life_expectancy(sc_makeham(40:110))
}
