% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{sc_lognormal_params}
\alias{sc_lognormal_params}
\title{Lognormal parameters from a mean and sd}
\usage{
sc_lognormal_params(mean, sd)
}
\arguments{
\item{mean}{Mean of the lognormal.}

\item{sd}{Standard deviation.}
}
\value{
\code{c(mu = , sigma = )}.
}
\description{
Lognormal parameters from a mean and sd
}
\examples{
sc_lognormal_params(1000, 500)
}
