% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{sc_note}
\alias{sc_note}
\title{Append a note to a table}
\usage{
sc_note(x, text)
}
\arguments{
\item{x}{A \code{scelo_table}.}

\item{text}{The note.}
}
\value{
The table, with the note appended.
}
\description{
Append a note to a table
}
