% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sc_sample}
\alias{sc_sample}
\title{Load a bundled sample}
\usage{
sc_sample(key = "claims", ...)
}
\arguments{
\item{key}{One of \code{claims}, \code{climate}, \code{dirty}, \code{wmtr-scenarios},
\code{lifelib-mp}, \code{workspace-demo}.}

\item{...}{Passed to \code{\link[=sc_load]{sc_load()}}.}
}
\value{
A data frame.
}
\description{
Load a bundled sample
}
\examples{
head(sc_sample("claims"))
}
