% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{sc_v}
\alias{sc_v}
\title{Discount factor}
\usage{
sc_v(i, t = 1)
}
\arguments{
\item{i}{Effective annual rate.}

\item{t}{Time(s) in years.}
}
\value{
Discount factor(s).
}
\description{
v^t = (1+i)^(−t).
}
\examples{
sc_v(0.05, 1:3)
}
